vc_viewer::handlers
Module pages

Page handlers

This module contains handlers for rendering the main pages of the application.

Structs

about

Handler for the about page

cookies

Handler for the cookie policy page

create_theme

Handler for the theme creation documentation page

docs

Handler for the documentation page

index_route

Handler for the index page

privacy

Handler for the privacy policy page

themes

Handler for the themes page

viewer

Handler for the viewer page

Functions

render_index

Generic index handler that can be used for both direct and catch-all routes